+++ src/main/java/org/springframework/beans/xml/XmlBeanDefinitionReader.java
@@ -1,0 +1,81 @@
+package org.springframework.beans.xml;
+
+import cn.hutool.core.util.StrUtil;
+import cn.hutool.core.util.XmlUtil;
+import org.springframework.beans.BeansException;
+import org.springframework.beans.PropertyValue;
+import org.springframework.beans.factory.config.BeanDefinition;
+import org.springframework.beans.factory.config.BeanReference;
+import org.springframework.beans.factory.support.BeanDefinitionRegistry;
+import org.springframework.beans.factory.support.DefaultListableBeanFactory;
+import org.w3c.dom.Document;
+import org.w3c.dom.Element;
+import org.w3c.dom.NodeList;
+
+import java.io.InputStream;
+
+/**
+ * @author: DoubleW2w
+ * @date: 2024/10/3
+ * @project: sbs-small-spring
+ */
+public class XmlBeanDefinitionReader {
+
+  private final DefaultListableBeanFactory defaultListableBeanFactory;
+
+  public XmlBeanDefinitionReader(DefaultListableBeanFactory defaultListableBeanFactory) {
+    this.defaultListableBeanFactory = defaultListableBeanFactory;
+  }
+
+  public void doLoadBeanDefinitions(InputStream inputStream) throws Exception {
+    Document doc = XmlUtil.readXML(inputStream);
+    Element root = doc.getDocumentElement();
+    NodeList childNodes = root.getChildNodes();
+    for (int i = 0; i < childNodes.getLength(); i++) {
+      // 判断元素
+      if (!(childNodes.item(i) instanceof Element)) continue;
+      // 判断对象
+      if (!"bean".equals(childNodes.item(i).getNodeName())) continue;
+
+      // 解析标签
+      Element bean = (Element) childNodes.item(i);
+      String id = bean.getAttribute("id");
+      String name = bean.getAttribute("name");
+      String className = bean.getAttribute("class");
+
+      // 获取 Class，方便获取类中的名称
+      Class<?> clazz = Class.forName(className);
+      // 优先级 id > name
+      String beanName = StrUtil.isNotBlank(id) ? id : name;
+      if (StrUtil.isNotBlank(beanName)) {
+        beanName = StrUtil.lowerFirst(clazz.getSimpleName());
+      }
+      // 定义Bean
+      BeanDefinition beanDefinition = new BeanDefinition(clazz);
+
+      // 读取属性并填充
+      NodeList beanChildNodes = bean.getChildNodes();
+      for (int j = 0; j < beanChildNodes.getLength(); j++) {
+        if (!(beanChildNodes.item(j) instanceof Element)) continue;
+        if (!"property".equals(beanChildNodes.item(j).getNodeName())) continue;
+        // 解析标签：property
+        Element property = (Element) beanChildNodes.item(j);
+        String attrName = property.getAttribute("name");
+        String attrValue = property.getAttribute("value");
+        String attrRef = property.getAttribute("ref");
+        // 获取属性值：引入对象、值对象
+        Object value = StrUtil.isNotBlank(attrRef) ? new BeanReference(attrRef) : attrValue;
+        // 创建属性信息
+        PropertyValue propertyValue = new PropertyValue(attrName, value);
+        beanDefinition.getPropertyValues().addPropertyValue(propertyValue);
+      }
+
+      // 如果存在相同的bean定义，就抛出异常
+      if (defaultListableBeanFactory.hasBeanDefinition(beanName)) {
+        throw new BeansException("Duplicate beanName[" + beanName + "] is not allowed");
+      }
+      // 注册 BeanDefinition
+      defaultListableBeanFactory.registerBeanDefinition(beanName, beanDefinition);
+    }
+  }
+}