+++ CHANGELOG.md
@@ -142,6 +142,8 @@ public class SimpleBeanContainerImplTest {
 
 ## Bean 实例化策略
 
+> 代码分支：[bean-instantiation-strategy](https://github.com/DoubleW2w/sbs-small-spring/tree/bean-instantiation-strategy)
+
 ### S
 
 在 Bean 的实例化过程，我们是通过反射获取 **无参构造函数** 去创建出来的对象。
@@ -258,6 +260,8 @@ public class CglibSubclassingInstantiationStrategy implements InstantiationStrat
 
 ## 为 Bean 对象注入属性
 
+> 代码分支：[bean-instantiation-strategy](https://github.com/DoubleW2w/sbs-small-spring/tree/bean-instantiation-strategy)
+
 ### S
 
 目前实现了无参构造、有参构造的实例化，但不能每次实例化 Bean，都要以构造函数的方式注入属性。
@@ -337,5 +341,133 @@ public class WorldService {
 
 ### R
 
-- 一开始是通过无参构造函数完成实例化Bean对象，后面增加了有参构造函数的方式，这也算是一种变相的注入属性。
+- 一开始是通过无参构造函数完成实例化 Bean 对象，后面增加了有参构造函数的方式，这也算是一种变相的注入属性。
 - 本节增加 Bean 属性的注入方式，通过反射的方式进行填充，不过这里使用的 hutool 的工具类来处理。
+
+## 从 Spring.xml 解析和注册 Bean 对象
+
+> 代码分支：[resource-load-registry-bean](https://github.com/DoubleW2w/sbs-small-spring/tree/resource-load-registry-bean)
+
+### S
+
+通过单元测试进行手动操作 Bean 对象的定义、注册和属性填充，以及最终获取对象调用方法，但在 Spring 中，一开始是通过 xml 方式进行配置的。
+
+### T
+
+将注册 Bean 的过程转移到 XML 配置文件中，完成 Spring 配置的读取、解析、注册。
+
+### A
+
+#### 最简单实现，通过 xml 标签的解析。
+
+```java
+public class XmlBeanDefinitionReader {
+
+  private final DefaultListableBeanFactory defaultListableBeanFactory;
+
+  public XmlBeanDefinitionReader(DefaultListableBeanFactory defaultListableBeanFactory) {
+    this.defaultListableBeanFactory = defaultListableBeanFactory;
+  }
+
+  public void doLoadBeanDefinitions(InputStream inputStream) throws Exception {
+    Document doc = XmlUtil.readXML(inputStream);
+    Element root = doc.getDocumentElement();
+    NodeList childNodes = root.getChildNodes();
+    for (int i = 0; i < childNodes.getLength(); i++) {
+      // 判断元素
+      if (!(childNodes.item(i) instanceof Element)) continue;
+      // 判断对象
+      if (!"bean".equals(childNodes.item(i).getNodeName())) continue;
+
+      // 解析标签
+      Element bean = (Element) childNodes.item(i);
+      String id = bean.getAttribute("id");
+      String name = bean.getAttribute("name");
+      String className = bean.getAttribute("class");
+
+      // 获取 Class，方便获取类中的名称
+      Class<?> clazz = Class.forName(className);
+      // 优先级 id > name
+      String beanName = StrUtil.isNotBlank(id) ? id : name;
+      if (StrUtil.isNotBlank(beanName)) {
+        beanName = StrUtil.lowerFirst(clazz.getSimpleName());
+      }
+      // 定义Bean
+      BeanDefinition beanDefinition = new BeanDefinition(clazz);
+
+      // 读取属性并填充
+      NodeList beanChildNodes = bean.getChildNodes();
+      for (int j = 0; j < beanChildNodes.getLength(); j++) {
+        if (!(beanChildNodes.item(j) instanceof Element)) continue;
+        if (!"property".equals(beanChildNodes.item(j).getNodeName())) continue;
+        // 解析标签：property
+        Element property = (Element) beanChildNodes.item(j);
+        String attrName = property.getAttribute("name");
+        String attrValue = property.getAttribute("value");
+        String attrRef = property.getAttribute("ref");
+        // 获取属性值：引入对象、值对象
+        Object value = StrUtil.isNotBlank(attrRef) ? new BeanReference(attrRef) : attrValue;
+        // 创建属性信息
+        PropertyValue propertyValue = new PropertyValue(attrName, value);
+        beanDefinition.getPropertyValues().addPropertyValue(propertyValue);
+      }
+
+      // 如果存在相同的bean定义，就抛出异常
+      if (defaultListableBeanFactory.hasBeanDefinition(beanName)) {
+        throw new BeansException("Duplicate beanName[" + beanName + "] is not allowed");
+      }
+      // 注册 BeanDefinition
+      defaultListableBeanFactory.registerBeanDefinition(beanName, beanDefinition);
+    }
+  }
+}
+```
+
+测试类
+
+```java
+public class ResourceLoadRegistryBeanTest {
+  @Test
+  public void test_xml() throws Exception {
+    // 1.初始化 BeanFactory
+    DefaultListableBeanFactory beanFactory = new DefaultListableBeanFactory();
+
+    // 2. 读取配置文件&注册Bean
+    XmlBeanDefinitionReader reader = new XmlBeanDefinitionReader(beanFactory);
+    reader.doLoadBeanDefinitions(ResourceUtil.getStream("spring-config.xml"));
+
+    // 3. 获取Bean对象调用方法
+    HelloService helloService = (HelloService) beanFactory.getBean("helloService");
+    String result = helloService.sayHello();
+    helloService.sayWorld();
+    System.out.println("测试结果：" + result);
+  }
+}
+```
+
+资源类的加载，我们通过 hutool 的 `ResourceUtil` 来解决，而 xml 内容的解析通过我们自定义的 `XmlBeanDefinitionReader` 来完成
+
+```xml
+<?xml version="1.0" encoding="UTF-8" ?>
+<!-- spring-config.xml -->
+<beans xmlns="http://www.springframework.org/schema/beans"
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xsi:schemaLocation="http://www.springframework.org/schema/beans
+       http://www.springframework.org/schema/beans/spring-beans.xsd">
+
+    <bean id="worldService" class="org.springframework.beans.factory.WorldService">
+    </bean>
+
+    <bean id="helloService" class="org.springframework.beans.factory.HelloService">
+        <property name="worldService" ref="worldService"/>
+        <property name="name" value="lisi"/>
+    </bean>
+</beans>
+```
+
+<img src="https://doublew2w-note-resource.oss-cn-hangzhou.aliyuncs.com/img/202410032016748.png"/>
+
+### R
+
+- 解析配置文件，注册 Bean 信息
+- 通过 Bean 工厂，获取 Bean，以及对应的调用方法。